// lib/types.ts
export type DriverData = {
    name: string;
    contactNumber: string;
    dlNo: string;
    dlExpiry: string;
    allowedVehicles: string;
    vehicleRegNum: string;
    vehicleType: string;
    make: string;
    model: string;
    capacity: string;
    insuranceExpiry: string;
    lastFC: string;
    fcExpiry: string;
    permit: string;
    puccExpiry: string;
};

export type IncidentReportData = {
    id: string;
    incidentDate: string;
    incidentTime: string;
    location: string;
    vehicleInvolved: string;
    driverInvolved: string;
    incidentType: string;
    severity: 'low' | 'medium' | 'high' | 'critical';
    description: string;
    photos: File[];
};

export type Rate = {
    id: string;
    vehicleType: string;
    category: 'passenger' | 'shipper' | 'equipment';
    percentPer: number;
    rate: number;
    vatPercent: number;
    vatAmount: number;
    damages: number;
    handling: number;
    waiting: number;
    halting: number;
};